nixos/qemu-vm: unify virtfs+drive setup logic
This was needed to be able to boot PowerNV VMs with the existing NixOS
VM PCI device setup. There should be no functional difference for
existing platforms.

PowerNV does not have a default PCIe layout, so every fsdev/drive has
to specify its pcie.x bus manually. Each device must be on its own bus,
so this had to be refactored to assign indices uniquely.

https://github.com/legoater/qemu/wiki/PowerNV#complex-pcie-configuration

Also, specifying the PCIe bus for a VirtFS device requires using the
longer -fsdev/-device form rather than the abbreviated -virtfs form.

https://wiki.qemu.org/Documentation/9psetup

It is possible to share the PCIe bus using a PCI bridge device, but this
does not avoid the uniqueness issue as the addresses on the PCI bridge must
still be unique.