nixos/qemu-flags: generalise qemuBinary for cross-compilation
In the general case, our command-line to invoke QEMU must consider
the host and guest platforms to know which flags will work.
For example, KVM can only be used for native virtualisation.

The qemuPkg arg only exists for backward compatibility. Ideally this would
simply be taken from the qemuHostPkgs package set, as is the default.