squash! wheel: Update FOD hash
wheel: Fix hash mismatch due to filesystem normalization

One file in the testdata has a name with accented unicode characters
that can be encoded differently depending on normalization. This causes
Nix to calculate a different hash for the tarball output depending on
whether or not and which unicode normal form the filesystem uses.

This is worked around by renaming the file to consist of unicode
characters that are unaffected by normalization. The file is renamed and
the test patched in the `extraPostFetch` phase of the the fetcher.