squash! sphinx: FOD hash changed, not sure why
sphinx: Patch source file for normalization

One of the test files has a different encoding depending on the
filesystem's normalization. By picking a character that does not have
multiple possible encodings we make the source indifferent to
normalization. This allows the FOD hash to match across platforms with
differently normalizing filesystems.